fixed bug in cropping of us rec data, changed intbb methods to use Vector3i
Former-commit-id: faa8d7e905b47b3c4520db5b28396c48b7cce4d4

+++ Code/Data/sscImage.cpp
@@ -667,7 +667,7 @@ void Image::mergevtkSettingsIntosscTransform()
 	Vector3D origin(mBaseImageData->GetOrigin());
 	Vector3D spacing(mBaseImageData->GetSpacing());
 	IntBoundingBox3D extent(mBaseImageData->GetExtent());
-	Vector3D extentShift = multiply_elems(extent.corner(0, 0, 0), spacing);
+	Vector3D extentShift = multiply_elems(extent.corner(0, 0, 0).cast<double>(), spacing);
 
 	vtkImageChangeInformationPtr info = vtkImageChangeInformationPtr::New();
 	info->SetInput(mBaseImageData);
